Move requirement status types to their own files, sync checklist on refresh
Split MobilePairingModel.RequirementStatus and its status mapping into
MobilePairingModel+RequirementStatus.swift and the shared badge view
into RequirementStatusBadge.swift, per the one-major-type-per-file
policy. refresh() now syncs the signed-in snapshot from the coordinator
before the bootstrap await and clears it when the coordinator is
missing, so a refresh triggered by sign-out never shows the prior
session's green signed-in row while loading.
